requirements.txt
pynetworktables==2019.0.2
git+https://github.com/dpallot/simple-websocket-server.git@master#egg=SimpleWebSocketServer
